Add annotation to dyn/stat properties and sketch; refactor properties; refactor FE events
