+++ data/small_example/small_example.json
@@ -1,4 +1,5 @@
 {
+  "annotation": "",
   "model": {
     "variables": [
       {
@@ -120,18 +121,21 @@
     {
       "id": "data_1",
       "name": "Generic dynamic property",
+      "annotation": "",
       "variant": "GenericDynProp",
       "formula": "3{x}:@{x}: ~v_1 & ~v_2 & v_3 & AG EF {x}"
     },
     {
       "id": "prior_knowledge",
       "name": "Generic dynamic property",
+      "annotation": "",
       "variant": "GenericDynProp",
       "formula": "3{a}: (3{b}: (3{c}: (@{c}: ((EF {a}) & (EF {b}) & (@{a}: AG EF {a}) & (@{b}: (AG EF {b} & ~EF {a}))))))"
     },
     {
       "id": "data_2",
       "name": "Generic dynamic property",
+      "annotation": "",
       "variant": "GenericDynProp",
       "formula": "3{y}:@{y}: v_1 & v_2 & ~v_3 & AG EF {y}"
     }
@@ -140,6 +144,7 @@
     {
       "id": "essentiality_v_3_v_1",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "v_3",
       "target": "v_1",
@@ -149,6 +154,7 @@
     {
       "id": "essentiality_v_2_v_3",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "v_2",
       "target": "v_3",
@@ -158,6 +164,7 @@
     {
       "id": "essentiality_v_1_v_3",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "v_1",
       "target": "v_3",
@@ -167,6 +174,7 @@
     {
       "id": "monotonicity_v_1_v_3",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "v_1",
       "target": "v_3",
@@ -176,6 +184,7 @@
     {
       "id": "monotonicity_v_3_v_1",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "v_3",
       "target": "v_1",
