+++ data/test_data/test_model.json
@@ -1,4 +1,5 @@
 {
+  "annotation": "",
   "model": {
     "variables": [
       {
@@ -148,6 +149,7 @@
     {
       "id": "essentiality_D_D",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "D",
       "target": "D",
@@ -157,6 +159,7 @@
     {
       "id": "essentiality_A_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "A",
       "target": "C",
@@ -166,6 +169,7 @@
     {
       "id": "monotonicity_B_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "B",
       "target": "C",
@@ -175,6 +179,7 @@
     {
       "id": "monotonicity_A_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "A",
       "target": "C",
@@ -184,6 +189,7 @@
     {
       "id": "essentiality_C_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "C",
       "target": "C",
@@ -193,6 +199,7 @@
     {
       "id": "essentiality_B_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "B",
       "target": "C",
@@ -202,6 +209,7 @@
     {
       "id": "monotonicity_A_B",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "A",
       "target": "B",
@@ -211,6 +219,7 @@
     {
       "id": "monotonicity_D_B",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "D",
       "target": "B",
@@ -220,6 +229,7 @@
     {
       "id": "essentiality_D_B",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "D",
       "target": "B",
@@ -229,6 +239,7 @@
     {
       "id": "monotonicity_C_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "C",
       "target": "C",
@@ -238,6 +249,7 @@
     {
       "id": "essentiality_A_B",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "A",
       "target": "B",
