+++ data/test_data/test_model_with_data.json
@@ -1,4 +1,5 @@
 {
+  "annotation": "",
   "model": {
     "variables": [
       {
@@ -197,6 +198,7 @@
     {
       "id": "essentiality_D_B",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "D",
       "target": "B",
@@ -206,6 +208,7 @@
     {
       "id": "essentiality_A_B",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "A",
       "target": "B",
@@ -215,6 +218,7 @@
     {
       "id": "monotonicity_A_B",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "A",
       "target": "B",
@@ -224,6 +228,7 @@
     {
       "id": "essentiality_D_D",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "D",
       "target": "D",
@@ -233,6 +238,7 @@
     {
       "id": "essentiality_B_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "B",
       "target": "C",
@@ -242,6 +248,7 @@
     {
       "id": "essentiality_C_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "C",
       "target": "C",
@@ -251,6 +258,7 @@
     {
       "id": "monotonicity_B_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "B",
       "target": "C",
@@ -260,6 +268,7 @@
     {
       "id": "monotonicity_C_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "C",
       "target": "C",
@@ -269,6 +278,7 @@
     {
       "id": "essentiality_A_C",
       "name": "Regulation essentiality property",
+      "annotation": "",
       "variant": "RegulationEssential",
       "input": "A",
       "target": "C",
@@ -278,6 +288,7 @@
     {
       "id": "monotonicity_A_C",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "A",
       "target": "C",
@@ -287,6 +298,7 @@
     {
       "id": "monotonicity_D_B",
       "name": "Regulation monotonicity property",
+      "annotation": "",
       "variant": "RegulationMonotonic",
       "input": "D",
       "target": "B",
