+++ src-tauri/src/analysis/_test_inference/_test_dynamic.rs
@@ -1,5 +1,6 @@
 use super::utils::load_test_model;
 use crate::analysis::_test_inference::utils::add_dyn_prop_and_infer;
+use crate::sketchbook::properties::shortcuts::*;
 use crate::sketchbook::properties::DynProperty;
 
 #[test]
@@ -9,42 +10,42 @@ fn inference_hctl() {
     let sketch = load_test_model();
     let formula = "3{x}: (3{y}: (@{x}: (AG~{y}) & (AG EF {x})) & (@{y}: AG EF {y}))";
     let id = "at_least_2_attrs";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 17);
 
     // Multiple fixed-points
     let sketch = load_test_model();
     let formula = "3{x}: (3{y}: (@{x}: ~{y} & (AX {x})) & (@{y}: AX {y}))";
     let id = "at_least_2_attrs";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 9);
 
     // No fixed-points
     let sketch = load_test_model();
     let formula = "~(3{x}: @{x}: AX {x})";
     let id = "at_least_2_attrs";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 20);
 
     // Has 1111 fixed point
     let sketch = load_test_model();
     let formula = "3{x}: @{x}: (A & B & C & D & AX {x})";
     let id = "has_1111_fixed_point";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 4);
 
     // State 1111 is part of an attractor
     let sketch = load_test_model();
     let formula = "3{x}: @{x}: (A & B & C & D & AG EF {x})";
     let id = "has_1111_in_attractor";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 19);
 
     // Has 111* minimal trap space
     let sketch = load_test_model();
     let formula = "V{x}: @{x}: (((A & B & C) => AG (A & B & C)) & ~ AX {x})";
     let id = "has_111X_mts";
-    let property = DynProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_hctl_prop(formula).unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 2);
 }
 
@@ -53,22 +54,22 @@ fn inference_hctl() {
 fn inference_template_attr_count() {
     let sketch = load_test_model();
     let id = "exactly_1_attr";
-    let property = DynProperty::try_mk_attractor_count(id, 1, 1).unwrap();
+    let property = DynProperty::try_mk_attractor_count(id, 1, 1, "").unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 15);
 
     let sketch = load_test_model();
     let id = "exactly_3_attr";
-    let property = DynProperty::try_mk_attractor_count(id, 3, 3).unwrap();
+    let property = DynProperty::try_mk_attractor_count(id, 3, 3, "").unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 1);
 
     let sketch = load_test_model();
     let id = "range_3_4_attr";
-    let property = DynProperty::try_mk_attractor_count(id, 3, 4).unwrap();
+    let property = DynProperty::try_mk_attractor_count(id, 3, 4, "").unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 2);
 
     let sketch = load_test_model();
     let id = "range_5_attr";
-    let property = DynProperty::try_mk_attractor_count(id, 5, 5).unwrap();
+    let property = DynProperty::try_mk_attractor_count(id, 5, 5, "").unwrap();
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 0);
 }
 
@@ -80,7 +81,7 @@ fn inference_template_fixed_point() {
     let id = "has_1111_fixed_point";
     let data_id = sketch.observations.get_dataset_id("data_fp").unwrap();
     let obs_id = sketch.observations.get_obs_id("data_fp", "ones").unwrap();
-    let property = DynProperty::mk_fixed_point(id, Some(data_id), Some(obs_id));
+    let property = DynProperty::mk_fixed_point(id, Some(data_id), Some(obs_id), "");
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 4);
 }
 
@@ -92,7 +93,7 @@ fn inference_template_attractor() {
     let id = "has_1111_in_attractor";
     let data_id = sketch.observations.get_dataset_id("data_fp").unwrap();
     let obs_id = sketch.observations.get_obs_id("data_fp", "ones").unwrap();
-    let property = DynProperty::mk_has_attractor(id, Some(data_id), Some(obs_id));
+    let property = DynProperty::mk_has_attractor(id, Some(data_id), Some(obs_id), "");
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 19);
 }
 
@@ -104,7 +105,7 @@ fn inference_template_trap_space() {
     let id = "has_111X_ts";
     let data_id = sketch.observations.get_dataset_id("data_mts").unwrap();
     let obs_id = sketch.observations.get_obs_id("data_mts", "abc").unwrap();
-    let property = DynProperty::mk_trap_space(id, Some(data_id), Some(obs_id), false, false);
+    let property = DynProperty::mk_trap_space(id, Some(data_id), Some(obs_id), false, false, "");
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 4);
 
     // Has 111* minimal trap space
@@ -112,6 +113,6 @@ fn inference_template_trap_space() {
     let id = "has_111X_mts";
     let data_id = sketch.observations.get_dataset_id("data_mts").unwrap();
     let obs_id = sketch.observations.get_obs_id("data_mts", "abc").unwrap();
-    let property = DynProperty::mk_trap_space(id, Some(data_id), Some(obs_id), true, true);
+    let property = DynProperty::mk_trap_space(id, Some(data_id), Some(obs_id), true, true, "");
     assert_eq!(add_dyn_prop_and_infer(sketch, property, id), 2);
 }
