+++ src-tauri/src/analysis/_test_inference/_test_static.rs
@@ -1,8 +1,8 @@
 use super::utils::load_test_model;
 use crate::analysis::_test_inference::utils::add_stat_prop_and_infer;
 use crate::sketchbook::model::{Essentiality, Monotonicity};
+use crate::sketchbook::properties::shortcuts::*;
 use crate::sketchbook::properties::StatProperty;
-use crate::sketchbook::stat_prop_utils::{mk_essentiality_prop, mk_monotonicity_prop};
 
 #[test]
 /// Test inference using the test model with added monotonicity properties in FOL.
@@ -11,21 +11,21 @@ fn inference_fol_monotonicity() {
     let sketch = load_test_model();
     let formula = "f_D(0) => f_D(1)";
     let id = "d_d_is_activation";
-    let property = StatProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_fol_prop(formula).unwrap();
     assert_eq!(add_stat_prop_and_infer(sketch, property, id), 16);
 
     // inhibition D -| D
     let sketch = load_test_model();
     let formula = "f_D(1) => f_D(0)";
     let id = "d_d_is_inhibition";
-    let property = StatProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_fol_prop(formula).unwrap();
     assert_eq!(add_stat_prop_and_infer(sketch, property, id), 16);
 
     // dual D -* D
     let sketch = load_test_model();
     let formula = "!(f_D(0) => f_D(1)) & !(f_D(1) => f_D(0))";
     let id = "d_d_is_dual";
-    let property = StatProperty::try_mk_generic(id, formula).unwrap();
+    let property = mk_fol_prop(formula).unwrap();
     assert_eq!(add_stat_prop_and_infer(sketch, property, id), 0);
 }
 
@@ -57,13 +57,13 @@ fn inference_fol_essentiality() {
     let sketch = load_test_model();
     let formula = "f_A(1) ^ f_A(0)";
     let id = "c_a_is_essential";
-    let property = StatProperty::try_mk_generic(id, formula).unwrap();
+    let property = StatProperty::try_mk_generic(id, formula, "").unwrap();
     assert_eq!(add_stat_prop_and_infer(sketch, property, id), 16);
 
     let sketch = load_test_model();
     let formula = "!(f_A(1) ^ f_A(0))";
     let id = "c_a_not_essential";
-    let property = StatProperty::try_mk_generic(id, formula).unwrap();
+    let property = StatProperty::try_mk_generic(id, formula, "").unwrap();
     assert_eq!(add_stat_prop_and_infer(sketch, property, id), 16);
 }
