+++ src-tauri/src/sketchbook/_sketch/_impl_export.rs
@@ -8,7 +8,7 @@ impl Sketch {
     ///
     /// See [SketchData::to_pretty_json_str] for details on the actual conversion to JSON.
     pub fn export_to_custom_json(&self, filepath: &str) -> Result<(), String> {
-        let sketch_data = SketchData::new(&self.model, &self.observations, &self.properties);
+        let sketch_data = SketchData::new_from_sketch(self);
         let mut file = File::create(filepath).map_err(|e| e.to_string())?;
         // write sketch in JSON to the file
         file.write_all(sketch_data.to_pretty_json_str().as_bytes())
