+++ src-tauri/src/sketchbook/model/_model_state/_impl_session_state/_events_regulations.rs
@@ -7,7 +7,8 @@ use crate::sketchbook::event_utils::{
 };
 use crate::sketchbook::ids::VarId;
 use crate::sketchbook::model::{Essentiality, ModelState, Monotonicity};
-use crate::sketchbook::stat_prop_utils::*;
+use crate::sketchbook::properties::shortcuts::*;
+use crate::sketchbook::properties::StatProperty;
 use crate::sketchbook::JsonSerde;
 
 /// Implementation for events related to `regulations` of the model.
@@ -58,14 +59,14 @@ impl ModelState {
         // events of adding the corresponding properties for monotonicity/essentiality in case it
         // is not unknown variant
         if reg_data.essential != Essentiality::Unknown {
-            let prop_id = get_essentiality_prop_id(&input_var, &target_var);
+            let prop_id = StatProperty::get_essentiality_prop_id(&input_var, &target_var);
             let prop = mk_essentiality_prop(&input_var, &target_var, reg_data.essential);
             let prop_payload = StatPropertyData::from_property(&prop_id, &prop).to_json_str();
             let prop_event = mk_stat_prop_event(&["add"], Some(&prop_payload));
             event_list.push(prop_event);
         }
         if reg_data.sign != Monotonicity::Unknown {
-            let prop_id = get_monotonicity_prop_id(&input_var, &target_var);
+            let prop_id = StatProperty::get_monotonicity_prop_id(&input_var, &target_var);
             let prop = mk_monotonicity_prop(&input_var, &target_var, reg_data.sign);
             let prop_payload = StatPropertyData::from_property(&prop_id, &prop).to_json_str();
             let prop_event = mk_stat_prop_event(&["add"], Some(&prop_payload));
@@ -139,13 +140,13 @@ impl ModelState {
             // case it is not unknown variant
             if *original_reg.get_essentiality() != Essentiality::Unknown {
                 // there is at max one essentiality property for a regulation
-                let prop_id = get_essentiality_prop_id(&regulator_id, &target_id);
+                let prop_id = StatProperty::get_essentiality_prop_id(&regulator_id, &target_id);
                 let prop_event = mk_stat_prop_event(&[prop_id.as_str(), "remove"], None);
                 event_list.push(prop_event);
             }
             if *original_reg.get_sign() != Monotonicity::Unknown {
                 // there is at max one monotonicity property for a regulation
-                let prop_id = get_monotonicity_prop_id(&regulator_id, &target_id);
+                let prop_id = StatProperty::get_monotonicity_prop_id(&regulator_id, &target_id);
                 let prop_event = mk_stat_prop_event(&[prop_id.as_str(), "remove"], None);
                 event_list.push(prop_event);
             }
@@ -192,7 +193,7 @@ impl ModelState {
 
             // events of modifying/adding/removing corresponding static property
             // note we have checked that `orig_sign` and `new_sign` are different
-            let prop_id = get_monotonicity_prop_id(&regulator_id, &target_id);
+            let prop_id = StatProperty::get_monotonicity_prop_id(&regulator_id, &target_id);
             if orig_sign == Monotonicity::Unknown {
                 // before there was no static prop, now we have to add it
                 let prop = mk_monotonicity_prop(&regulator_id, &target_id, new_sign);
@@ -261,7 +262,7 @@ impl ModelState {
 
             // events of modifying/adding/removing corresponding static property
             // note we have checked that `orig_essentiality` and `new_essentiality` are different
-            let prop_id = get_essentiality_prop_id(&regulator_id, &target_id);
+            let prop_id = StatProperty::get_essentiality_prop_id(&regulator_id, &target_id);
             if orig_essentiality == Essentiality::Unknown {
                 // before there was no static prop, now we have to add it
                 let prop = mk_essentiality_prop(&regulator_id, &target_id, new_essentiality);
